[chore] Rename Python package namespace to opencode_a2a_serve (#59)
