+++ src/opencode_a2a_serve/agent.py
