+++ src/opencode_a2a_serve/jsonrpc_ext.py
