+++ tests/test_directory_validation.py
@@ -5,9 +5,9 @@ import pytest
 from a2a.server.agent_execution import RequestContext
 from a2a.server.events.event_queue import EventQueue
 
-from opencode_a2a.agent import OpencodeAgentExecutor
-from opencode_a2a.config import Settings
-from opencode_a2a.opencode_client import OpencodeClient
+from opencode_a2a_serve.agent import OpencodeAgentExecutor
+from opencode_a2a_serve.config import Settings
+from opencode_a2a_serve.opencode_client import OpencodeClient
 
 
 @pytest.fixture
