+++ tests/test_opencode_session_extension.py
@@ -3,7 +3,7 @@ import logging
 import httpx
 import pytest
 
-from opencode_a2a.config import Settings
+from opencode_a2a_serve.config import Settings
 
 
 class DummyOpencodeClient:
@@ -61,7 +61,7 @@ def _settings(*, token: str, log_payloads: bool) -> Settings:
 
 @pytest.mark.asyncio
 async def test_session_query_extension_requires_bearer_token(monkeypatch):
-    import opencode_a2a.app as app_module
+    import opencode_a2a_serve.app as app_module
 
     monkeypatch.setattr(app_module, "OpencodeClient", DummyOpencodeClient)
     app = app_module.create_app(_settings(token="t-1", log_payloads=False))
@@ -88,7 +88,7 @@ async def test_session_query_extension_requires_bearer_token(monkeypatch):
 
 @pytest.mark.asyncio
 async def test_session_query_extension_returns_jsonrpc_result(monkeypatch):
-    import opencode_a2a.app as app_module
+    import opencode_a2a_serve.app as app_module
 
     dummy = DummyOpencodeClient(_settings(token="t-1", log_payloads=False))
     monkeypatch.setattr(app_module, "OpencodeClient", lambda _settings: dummy)
@@ -147,7 +147,7 @@ async def test_session_query_extension_returns_jsonrpc_result(monkeypatch):
 
 @pytest.mark.asyncio
 async def test_session_query_extension_items_is_always_array(monkeypatch):
-    import opencode_a2a.app as app_module
+    import opencode_a2a_serve.app as app_module
 
     class WeirdPayloadClient(DummyOpencodeClient):
         def __init__(self, _settings: Settings) -> None:
@@ -177,7 +177,7 @@ async def test_session_query_extension_items_is_always_array(monkeypatch):
 
 @pytest.mark.asyncio
 async def test_session_query_extension_session_title_is_extracted_or_placeholder(monkeypatch):
-    import opencode_a2a.app as app_module
+    import opencode_a2a_serve.app as app_module
 
     class TitlePayloadClient(DummyOpencodeClient):
         def __init__(self, _settings: Settings) -> None:
@@ -203,7 +203,7 @@ async def test_session_query_extension_session_title_is_extracted_or_placeholder
 
 @pytest.mark.asyncio
 async def test_session_query_extension_message_role_and_id_from_info(monkeypatch):
-    import opencode_a2a.app as app_module
+    import opencode_a2a_serve.app as app_module
 
     class InfoRoleClient(DummyOpencodeClient):
         def __init__(self, _settings: Settings) -> None:
@@ -242,7 +242,7 @@ async def test_session_query_extension_message_role_and_id_from_info(monkeypatch
 
 @pytest.mark.asyncio
 async def test_session_query_extension_accepts_top_level_list_payload(monkeypatch):
-    import opencode_a2a.app as app_module
+    import opencode_a2a_serve.app as app_module
 
     class ListPayloadClient(DummyOpencodeClient):
         def __init__(self, _settings: Settings) -> None:
@@ -281,7 +281,7 @@ async def test_session_query_extension_accepts_top_level_list_payload(monkeypatc
 
 @pytest.mark.asyncio
 async def test_session_query_extension_accepts_alternative_list_keys(monkeypatch):
-    import opencode_a2a.app as app_module
+    import opencode_a2a_serve.app as app_module
 
     class AltKeyPayloadClient(DummyOpencodeClient):
         def __init__(self, _settings: Settings) -> None:
@@ -320,7 +320,7 @@ async def test_session_query_extension_accepts_alternative_list_keys(monkeypatch
 
 @pytest.mark.asyncio
 async def test_session_query_extension_rejects_cursor_limit(monkeypatch):
-    import opencode_a2a.app as app_module
+    import opencode_a2a_serve.app as app_module
 
     dummy = DummyOpencodeClient(_settings(token="t-1", log_payloads=False))
     monkeypatch.setattr(app_module, "OpencodeClient", lambda _settings: dummy)
@@ -348,7 +348,7 @@ async def test_session_query_extension_rejects_cursor_limit(monkeypatch):
 
 @pytest.mark.asyncio
 async def test_session_query_extension_rejects_size_over_max(monkeypatch):
-    import opencode_a2a.app as app_module
+    import opencode_a2a_serve.app as app_module
 
     dummy = DummyOpencodeClient(_settings(token="t-1", log_payloads=False))
     monkeypatch.setattr(app_module, "OpencodeClient", lambda _settings: dummy)
@@ -375,7 +375,7 @@ async def test_session_query_extension_rejects_size_over_max(monkeypatch):
 
 @pytest.mark.asyncio
 async def test_session_query_extension_maps_404_to_session_not_found(monkeypatch):
-    import opencode_a2a.app as app_module
+    import opencode_a2a_serve.app as app_module
 
     class NotFoundOpencodeClient(DummyOpencodeClient):
         async def list_messages(self, session_id: str, *, params=None):
@@ -408,10 +408,10 @@ async def test_session_query_extension_maps_404_to_session_not_found(monkeypatch
 
 @pytest.mark.asyncio
 async def test_session_query_extension_does_not_log_response_bodies(monkeypatch, caplog):
-    import opencode_a2a.app as app_module
+    import opencode_a2a_serve.app as app_module
 
     monkeypatch.setattr(app_module, "OpencodeClient", DummyOpencodeClient)
-    caplog.set_level(logging.DEBUG, logger="opencode_a2a.app")
+    caplog.set_level(logging.DEBUG, logger="opencode_a2a_serve.app")
 
     app = app_module.create_app(_settings(token="t-1", log_payloads=True))
