+++ tests/test_settings.py
@@ -4,7 +4,7 @@ from unittest import mock
 import pytest
 from pydantic import ValidationError
 
-from opencode_a2a.config import Settings
+from opencode_a2a_serve.config import Settings
 
 
 def test_settings_missing_required():
